refactor(j2cl): split the selected-wave projector test into 9 per-scenario classes (#1270)
Second half of the #1270 test split (the compose split landed in #1318).
J2clSelectedWaveProjectorTest was a ~4.1k-line monolith.

- Introduced abstract J2clSelectedWaveProjectorTestSupport (package-private
  shared update/fragment/document builders + constants).
- Split the 110 tests by projection scenario into 9 classes, each extending the
  base and each < 800 LOC (max 724): ReadState, ViewportProjection,
  DocumentFallback, InteractionMetadata, ViewportWindow, DocumentViewportMerge,
  WriteSession, ReadBlipMetadata, LockAndManifest.
- Deleted the original monster file; documented the split in the verification
  matrix.

All 110 tests compile and pass (surefire: 9 classes, 110 run, 0 failures).
